+++ lua/haskell-tools/init.lua
@@ -1,6 +1,7 @@
 local M = {
   config = nil,
   lsp = nil,
+  hoogle = nil,
 }
 
 function M.setup(opts)
@@ -8,9 +9,12 @@ function M.setup(opts)
   M.config = config
   local lsp = require('haskell-tools.lsp')
   M.lsp = lsp
+  local hoogle = require('haskell-tools.hoogle')
+  M.hoogle = hoogle
 
   config.setup(opts)
   lsp.setup()
+  hoogle.setup()
 end
 
 return M